[AIRFLOW-5300] Move GcsToService operators to core
For more information check AIP-21.
