+++ tests/operators/test_gcs_to_s3.py
@@ -19,7 +19,7 @@
 
 import unittest
 
-from airflow.contrib.operators.gcs_to_s3 import GoogleCloudStorageToS3Operator
+from airflow.operators.gcs_to_s3 import GoogleCloudStorageToS3Operator
 from airflow.hooks.S3_hook import S3Hook
 from tests.compat import mock
 
@@ -41,7 +41,7 @@ class TestGoogleCloudStorageToS3Operator(unittest.TestCase):
     # Test1: incremental behaviour (just some files missing)
     @mock_s3
     @mock.patch('airflow.contrib.operators.gcs_list_operator.GoogleCloudStorageHook')
-    @mock.patch('airflow.contrib.operators.gcs_to_s3.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.gcs_to_s3.GoogleCloudStorageHook')
     def test_execute_incremental(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -71,7 +71,7 @@ class TestGoogleCloudStorageToS3Operator(unittest.TestCase):
     # Test2: All the files are already in origin and destination without replace
     @mock_s3
     @mock.patch('airflow.contrib.operators.gcs_list_operator.GoogleCloudStorageHook')
-    @mock.patch('airflow.contrib.operators.gcs_to_s3.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.gcs_to_s3.GoogleCloudStorageHook')
     def test_execute_without_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -101,7 +101,7 @@ class TestGoogleCloudStorageToS3Operator(unittest.TestCase):
     # Test3: There are no files in destination bucket
     @mock_s3
     @mock.patch('airflow.contrib.operators.gcs_list_operator.GoogleCloudStorageHook')
-    @mock.patch('airflow.contrib.operators.gcs_to_s3.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.gcs_to_s3.GoogleCloudStorageHook')
     def test_execute(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -130,7 +130,7 @@ class TestGoogleCloudStorageToS3Operator(unittest.TestCase):
     # Test4: Destination and Origin are in sync but replace all files in destination
     @mock_s3
     @mock.patch('airflow.contrib.operators.gcs_list_operator.GoogleCloudStorageHook')
-    @mock.patch('airflow.contrib.operators.gcs_to_s3.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.gcs_to_s3.GoogleCloudStorageHook')
     def test_execute_with_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -160,7 +160,7 @@ class TestGoogleCloudStorageToS3Operator(unittest.TestCase):
     # Test5: Incremental sync with replace
     @mock_s3
     @mock.patch('airflow.contrib.operators.gcs_list_operator.GoogleCloudStorageHook')
-    @mock.patch('airflow.contrib.operators.gcs_to_s3.GoogleCloudStorageHook')
+    @mock.patch('airflow.operators.gcs_to_s3.GoogleCloudStorageHook')
     def test_execute_incremental_with_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
